% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutAmplicon.R
\name{cutAmplicon}
\alias{cutAmplicon}
\title{Truncate the amplicons resulting from insilicoPCR}
\usage{
cutAmplicon(amplicon, cut, outpath)
}
\arguments{
\item{amplicon}{(mandatory) DNAStringSet object: the amplicons or sequences to cut.}

\item{cut}{(mandatory) a numeric vector containing the start and the end (in bp) to define where to cut. Example: c(200,800) to keep only the region from 200bp to 800bp.}

\item{outpath}{(optional) the path to the FOLDER in which the function outputs will be stored.}
}
\value{
a DNAStringSet object with the remaining parts of the amplicons or sequences.
}
\description{
The goal of this function is simply to gather one part of the amplicon (or sequences) from a location to another (in bp).
}
\note{
To save the outputs (as .Rdata) in a given folder (here "JellyOutputs")
add the following argument: outpath = "path_to/JellyOutputs"
}
\examples{
\dontrun{
JellyFrag_100_to_500 <- OligoTestR(JellyDNA, cut = c(100,500))
}
}
