% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DegeScore.R
\name{DegeScore}
\alias{DegeScore}
\title{Calculate Degenerescence Score}
\usage{
DegeScore(x)
}
\arguments{
\item{x}{(mandatory) an oligonucleotide or sequence (character)}
}
\value{
The degenerescence score (numeric) associated to the oligonucleotide/sequences provided as x.
}
\description{
This function allows to calculate the degenerescence score for a given oligonucleotide or sequence.
(W, S, R, Y, K, M have a 2-deg. score ;
B, D, H, V have a 3-deg. score ;
N has a 4-deg. score).
}
