% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IUPAC_COMPLETE_MAP_OBJECT.R
\docType{data}
\name{IUPAC_COMPLETE_MAP}
\alias{IUPAC_COMPLETE_MAP}
\title{IUPAC_COMPLETE_MAP}
\format{
An object of class \code{character} of length 65.
}
\usage{
Especially called in BuildConsensus function
}
\description{
Creates a complete mapping of all nucleotides, as IUPAC_CODE_MAP from Biostrings, but including gaps as "-"
and specifications to attribute degenerated code to nucleotides
}
\keyword{datasets}
