% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFASTAasDNA.R
\name{loadFASTAasDNA}
\alias{loadFASTAasDNA}
\title{FASTA loader}
\usage{
loadFASTAasDNA(fasta)
}
\arguments{
\item{fasta}{The path of the FASTA file to load in R.}
}
\value{
A DNAStringSet object (list of DNA sequences).
}
\description{
Load a FASTA file and convert it to a DNA list (DNAStringSet object from Biostrings package).
}
\examples{
\dontrun{
JellyDNA <- loadFASTAasDNA("training_files/Cubozoa_and_Scyphozoa.fasta")
}
}
