% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OligoFindR.R
\name{OligoFindR}
\alias{OligoFindR}
\title{Find and extract oligonucleotides}
\usage{
OligoFindR(
  ConsensusClust,
  PriMin = 18,
  PriMax = 22,
  maxDeg = 2,
  GCrange = c(40, 60),
  outpath
)
}
\arguments{
\item{ConsensusClust}{(mandatory) it should be the output from BuildConsensus() function, a list of consensus sequences.}

\item{PriMin}{(mandatory, set to 18 by default) the minimum size (in bp) expected for the primers.}

\item{PriMax}{(mandatory, set to 22 by default) the maximum size (in bp) expected for the primers.}

\item{maxDeg}{(mandatory, set to 2 by default) the maximum of degenerescence allowed for the primers.
(The degenerescence score of W -which is A or T- is 2 ; the one of B -which can be T or C or G- is 3 ; N is 4 etc.).}

\item{GCrange}{(mandatory, set to c(40,60) by default) the range (in \%) of \%GC content within the oligo.}

\item{outpath}{(optional) the path to the FOLDER in which the function outputs will be stored.}
}
\value{
A dataframe with the oligonucleotides fulfilling the given parameters and kept from each clusters.
}
\description{
This function allows to look for conserved oligonucleotides in each consensus sequences according to
the parameters given by the user (see args).
}
\note{
Example made using default settings
To save the dataframe (as .Rdata) in a given folder (here "JellyOutputs")
add the following argument: outpath = "path_to/JellyOutputs"
}
\examples{
\dontrun{
JellyOligoFound <- OligoFindR(JellyConsensus, PriMin=18, PriMax=22, maxDeg=0, GCrange=c(40,60))
}
}
