% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OligoMatchR.R
\name{OligoMatchR}
\alias{OligoMatchR}
\title{Look for suitable oligonucleotides to create primers pairs}
\usage{
OligoMatchR(OligoChecked, ampSize, target, GCtail = TRUE, outpath)
}
\arguments{
\item{OligoChecked}{(mandatory) it must be the output from OligoCheckR() function, the dataframe with the oligonucleotides kept and their full characteristics.}

\item{ampSize}{(mandatory) a numeric vector is expected, with the minimum and maximum size range of the amplicon. Example: ampSize = c(400,600) to select amplicon from 400pb to 600pb (primers included in the size).}

\item{target}{(optional) a numeric vector is expected, if you want to look for primers couples only in a particular region within the sequences.
You should indicate the values as a numeric vector: (for example c(400,800) to target ONLY a variable region located around 400 and 800pb).}

\item{GCtail}{(mandatory, default: TRUE) if TRUE then the oligonucleotides presenting 3 or more G, C (and other bases corresponding) on their last nucleotides at the 3' tail will be removed from the analysis.}

\item{outpath}{(optional) the path to the FOLDER in which the function outputs will be stored.}
}
\value{
Another dataframe of the matching oligonucleotides couples with the given parameters. In this one, all the nucleotides are given in the 5'->3' orientation.
}
\description{
This function allows to check if, among the oligonucleotides previously kept, some of them could constitute pairs to
amplify gene fragments suitable with the size expected for the final amplicons.
}
\note{
Example made using default settings
To save the dataframe (as .Rdata) in a given folder (here "JellyOutputs")
add the following argument: outpath = "path_to/JellyOutputs"
}
\examples{
\dontrun{
JellyOligoMatched <- OligoMatchR(JellyOligoChecked, ampSize = c(300,500), GCtail = TRUE)
}
}
