% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OligoSpecR.R
\name{OligoSpecR}
\alias{OligoSpecR}
\title{Check the primers specificity to the target sequences}
\usage{
OligoSpecR(PrimersCouples, TargetSeq, AvoidSeq, mmCalc = 2, outpath)
}
\arguments{
\item{PrimersCouples}{(mandatory) a set of different primers (output from the 'OligoMatchR' or 'OligoTestR' function) OR a vector composed by the 2 primers (in 5'->3' conventional orientation) such as c("FORWARD","REVERSE").}

\item{TargetSeq}{(mandatory) a DNAStringSet object (such as the output from 'loadFASTAasDNA' function) containing the sequences to target by the amplification with the primers couple(s).}

\item{AvoidSeq}{(mandatory) a DNAStringSet object (such as the output from 'loadFASTAasDNA' function) containing the sequences for which the amplification by the primers couple(s) should be avoided.}

\item{mmCalc}{(mandatory, set to 2 by default) the number of mismatches used in the primers used in the in silico PCR process.}

\item{outpath}{(optional) the path to the FOLDER in which the function outputs will be stored.}
}
\value{
A dataframe showing the number of sequences amplified from each dataset and a specificity score of the primers couples.
}
\description{
This function allows to check the specificity of a set of primers couples to a selection of sequences.
The specificity will be defined according to the ability to amplify -in silico- a target dataset of sequences
while failing to amplify another dataset of sequences to exclude.
}
\note{
Example made using default settings and using only the first 10 primers obtained from OligoTestR function.
To save the different consensus (as .Rdata) in a given folder (here "JellyOutputs")
add the following argument: outpath = "path_to/JellyOutputs"
}
\examples{
\dontrun{
OligoSpecR(OligoResults[1:10,], JellyDNA, NOJellyDNA, mmCalc = 2)
}
}
