% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OligoTestR.R
\name{OligoTestR}
\alias{OligoTestR}
\title{Test a set of different primers on a dataset}
\usage{
OligoTestR(DNAList, PrimersCouples, mmCalc = 2, outpath)
}
\arguments{
\item{DNAList}{(mandatory) a DNAStringSet object, such as the output from 'loadFASTAasDNA' function.}

\item{PrimersCouples}{(mandatory) a set of different primers (output from the 'OligoMatchR' function) OR a vector composed by the 2 primers (in 5'->3' conventional orientation) such as c("FORWARD","REVERSE").}

\item{mmCalc}{(mandatory, set to 2 by default) the number of mismatches used in the primers scoring calculation and the amplicon resolution showed.}

\item{outpath}{(optional) the path to the FOLDER in which the function outputs will be stored.}
}
\value{
A dataframe of the matching oligonucleotides couples with the given parameters. In this one, all the nucleotides are given in the 5'->3' orientation.
}
\description{
This function allows to test a set of different primers on a dataset of DNA sequences in order to sort them by their amplification ability.
}
\note{
Example made using default settings
To save the dataframe (as .Rdata) in a given folder (here "JellyOutputs")
add the following argument: outpath = "path_to/JellyOutputs"
}
\examples{
\dontrun{
OligoTestR(JellyDNA, JellyOligoMatched, mmCalc = 2)
}
}
