% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revCompCase.R
\name{revCompCase}
\alias{revCompCase}
\title{Reverse complement (case sensitive)}
\usage{
revCompCase(seq, class = "vector")
}
\arguments{
\item{seq}{(mandatory) the sequence or oligonucleotides to set as reverse complement (character or DNAString object).}

\item{class}{(default: vector) the class of the reverse complement returned by the function (could be "vector" or "DNAString").}
}
\value{
the reverse complement as a vector (chain of characters) or a DNAString object.
}
\description{
This function allows to create the reverse complement of an oligonucleotides keeping the characters in upper and lowercases (as well compatible with gaps and degenerated bases).
}
