% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShuffleAndAlign.R
\name{ShuffleAndAlign}
\alias{ShuffleAndAlign}
\title{Shuffle the sequences, divide them into clusters and align them}
\usage{
ShuffleAndAlign(DNAList, nseq = 10, ite = 10, outpath)
}
\arguments{
\item{DNAList}{(mandatory) should be a DNAStringSet object (from loadFASTAasDNA function).}

\item{nseq}{(mandatory, set to 10 by default) the number of sequences in each cluster.}

\item{ite}{(mandatory, set to 10 by default) the number of iterations of the shuffling process.
Set 1 for no iteration, higher is the iteration number, stronger will be the results but higher will be the processing time.}

\item{outpath}{(optional) the path to the FOLDER in which the function outputs will be stored.}
}
\value{
A list of clusters of aligned sequences (each cluster are MsaDNAMulipleAlignment object)
}
\description{
This function enables to create random clusters of sequences from the dataset and
align the sequences within each cluster according to the parameters given by the user.
To increase the robustness of the future results, it is also possible to iterate the creation of random cluster of sequences.
}
\note{
This function may take some time to process, especially for long sequences, big dataset or high number of iterations.

Example made using default settings
To save the alignments list (as .Rdata) in a given folder (here "JellyOutputs")
add the following argument: outpath = "path_to/JellyOutputs"
}
\examples{
\dontrun{
JellyAlignments <- ShuffleAndAlign(JellyDNA, nseq = 10, ite = 10)
}
}
